[feat] UserMapper.xml 구현 - UserMapper.xml 현재 UserMapper 인터페이스에 존재하는 만큼 구현

+++ src/main/java/com/honeyboard/api/user/model/mapper/UserMapper.java
@@ -5,11 +5,11 @@ import com.honeyboard.api.user.model.User;
 public interface UserMapper {
 	
 	Integer insertUser(User user);
-	User selectUserById(long userId);
+	User selectUserById(int userId);
 	User selectUserByEmail(String email);
 	Integer updatePassword(User user);
 	Integer updateUser(User user);
-	Integer updateUserCompletionStatus(long userId);
+	Integer updateUserCompletionStatus(int userId);
 	Integer selectExistedEmail(String email);
 
 }
