+++ src/main/java/com/honeyboard/api/user/model/service/UserService.java
@@ -5,11 +5,11 @@ import com.honeyboard.api.user.model.User;
 public interface UserService {
 
 	Integer saveUser(User user);
-	User getUser(long userId);
+	User getUser(int userId);
 	User getUserByEmail(String email);
 	Integer updatePassword(User user);
 	Integer updateUser(User user);
-	Integer updateUserCompletionStatus(long userId);
+	Integer updateUserCompletionStatus(int userId);
 	Integer existsByEmail(String email);
 	
 }
